fastapi authenticate, cors

+++ routes/events.py
@@ -1,6 +1,7 @@
 from beanie import PydanticObjectId
-from fastapi import APIRouter, Body, HTTPException, status
+from fastapi import APIRouter, Body, Depends, HTTPException, status
 from database.connection import Database
+from auth.authenticate import authenticate
 
 from models.events import Event, EventUpdate
 from typing import List
@@ -12,6 +13,7 @@ event_router = APIRouter(
 )
 
 events = []
+
 
 # 도든 이벤트 추출
 @event_router.get("/", response_model=List[Event])
@@ -30,16 +32,25 @@ async def retrieve_event(id: PydanticObjectId) -> Event:
         )
     return event
 
+
 # 이벤트 생성
 @event_router.post("/new")
-async def create_event(body: Event = Body(...)) -> dict:
+async def create_event(body: Event, user: str = Depends(authenticate)) -> dict:
+    body.creator = user
     await event_database.save(body)
     return {
         "message": "Event created successfully."
     }
 
+
 @event_router.put("/{id}", response_model=Event)
-async def update_event(id: PydanticObjectId, body: EventUpdate) -> Event:
+async def update_event(id: PydanticObjectId, body: EventUpdate, user: str = Depends(authenticate)) -> Event:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Operation not allowed"
+        )
     updated_event = await event_database.update(id, body)
     if not updated_event:
         raise HTTPException(
@@ -48,9 +59,16 @@ async def update_event(id: PydanticObjectId, body: EventUpdate) -> Event:
         )
     return updated_event
 
+
 # 데이터베이스에 있는 단일 이벤트 삭제
 @event_router.delete("/{id}")
-async def delete_event(id: PydanticObjectId) -> dict:
+async def delete_event(id: PydanticObjectId, user: str = Depends(authenticate)) -> dict:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_NOT_FOUND,
+            detail="Operation not allowed"
+        )
     event = await event_database.delete(id)
     if not event:
         raise HTTPException(
@@ -60,6 +78,7 @@ async def delete_event(id: PydanticObjectId) -> dict:
     return {
         "message": "Event deleted successfully."
     }
+
 
 # 전체 이벤트 삭제
 @event_router.delete("/")
